fix(migration): ajusta tipo da coluna valor em compra para NUMERIC(10,2)

+++ src/main/resources/db/migration/V1__create_tables.sql
@@ -15,7 +15,7 @@ CREATE TABLE compra
 (
     id                BIGSERIAL PRIMARY KEY,
     descricao         VARCHAR(255),
-    valor             DOUBLE PRECISION,
+    valor             NUMERIC(10,2),
     data              DATE,
     categoria_id      BIGINT,
     cartao_credito_id BIGINT,
